Fix Halon vsi to generate random system-mac
Halon needs to generate a different mac for each instance
of the switch. Without this, all the switches on a mininet
topology will end up with the same mac address. And since
we are using the system-mac as the interface mac, all the
mac addresses for all the switches are the same.

Change-Id: Ia896401aa6eac5e841f135a92451e62a2e1d5d3e
Signed-off-by: Rahul Rajagopal <rahulr@hp.com>