Change the seed for random number generation to have unique MAC addresses for the interfaces.
TG-1251

Time in seconds was used as the seed for srand() and sometimes this could be
the same on switches that came up at the same time(within a second). Changed
the seed to have the lower bits of the time in nano seconds.

Tags : chg, dev

Change-Id: I5a2fbde3720771bc01ddb84e3cc3b7af50235107
Signed-off-by: Sudhakar Mohan <sudhakar.mohan@hpe.com>
(cherry picked from commit fc2d85607a08b365f7bf6ce323c646e8a2ba3915)